// frontend/ui-kit/components/molecule/ShoppingListItemCard/ShoppingListItemCard.module.scss.d.ts
declare const styles: {
  shoppingListItemCard: string;
  shoppingListItemCard__name: string;
  shoppingListItemCard__quantity: string;
  shoppingListItemCard__price: string;
  shoppingListItemCard__category: string;
  shoppingListItemCard__focusBtn: string;
  shoppingListItemCard__focusBtnActive: string;
  shoppingListItemCard__editBtn: string;
  shoppingListItemCard__deleteBtn: string;
  readonly "shoppingListItemCard__name--focused": string;
  readonly "shoppingListItemCard__name--bought": string;
};
export default styles; 